extract the endpoint
endpoint extraction
javascript confirmation window while deletion

+++ ppm-tool-react-client/src/actions/projectActions.js
@@ -42,7 +42,12 @@ export const getProject = (id, history) => async dispatch => {
 };
 
 export const deleteProject = id => async dispatch => {
-  const res = await axios.delete(`http://localhost:8090/api/project/${id}`);
+  if (
+    window.confirm(
+      "Are you sure? This will delete the project and all the data related to it"
+    )
+  )
+    await axios.delete(`http://localhost:8090/api/project/${id}`);
   dispatch({
     type: DELETE_PROJECT, // this is calling to projectReducer with id parameter
     payload: id
